Styling and input rewrites
This comments out the websocket code because the server can't be run on
anything but a pi, since it depends on the GPIO code.

This makes a lot of changes to the input events to make them properly
work on mobile. Just using the click events doesn't seem to behave
properly when clicking and holding, or dragging on mobile, so we need
actual touch events for this. In addition, there's no touchleave event
or similar, so we're forced to emulate this.

The markup and style changes are the make the buttons large enough to
view on mobile. Currently the :active styling is somewhat misleading,
because we deactivate it when we drag off, while the active state
remains as long as the touch/click is held down.