Make the active highlights more accurate
The :active CSS class wasn't accurate--when you click something it
starts, and then it stays on as long as that click is held. This doesn't
match the way we handle button activation. If you drag off the button
while holding it, we end the held state. Using a CSS class gives us more
control and lets us force it to match.